Ratatoskr2 restructured with yeoman-angular
just the angular frontend. This commit serves to see the difference
between the boilerplate and the content of the angular app. ng-constant
is used to load env variables into a config file instead of angular-
deffered-bootstrap loading it from the backend.
Huginn local application.properties overrides the one packaged in
resources.